move code from Intrinsix to jvm/Intrinsics
fixes #2023
