+++ lib/fuzion/sys/env_vars.fz
@@ -32,21 +32,21 @@ module env_vars is
   #
   # NOTE: parameter s must be 0-terminated char *
   #
-  private has0(s Any) bool is intrinsic
+  private has0(s fuzion.sys.Pointer) bool is intrinsic
 
 
   # intrinsic to get env var with given name
   #
   # NOTE: parameter s must be 0-terminated char *
   #
-  private get0(s Any) String
+  private get0(s fuzion.sys.Pointer) String
   is intrinsic
 
   # intrinsic to set env var
-  private set0(s Any, t Any) bool is intrinsic
+  private set0(s fuzion.sys.Pointer, t fuzion.sys.Pointer) bool is intrinsic
 
   # intrinsic to unset env var
-  private unset0(s Any) bool is intrinsic
+  private unset0(s fuzion.sys.Pointer) bool is intrinsic
 
 
   # check if env var with given name exists
