+++ lib/fuzion/sys/net.fz
@@ -47,7 +47,7 @@ module net is
   # 0  => arr_result[0] is the socket descriptor
   # -1 => arr_result[0] is an error number
   #
-  bind0   (family, socket_type, protocol i32, host Any, port Any, arr_result Any) i32 is intrinsic
+  bind0   (family, socket_type, protocol i32, host fuzion.sys.Pointer, port fuzion.sys.Pointer, arr_result fuzion.sys.Pointer) i32 is intrinsic
 
 
   # bind a name to a newly created socket, wrapper for bind0 intrinsic
@@ -78,7 +78,7 @@ module net is
   # true  => arr_result[0] is the socket descriptor
   # false => arr_result[0] is an error number
   #
-  accept (sd i64, arr_result Any) bool is intrinsic
+  accept (sd i64, arr_result fuzion.sys.Pointer) bool is intrinsic
 
 
   # accept a new connection for given socket descriptor, wrapper for accept intrinsic.
@@ -93,7 +93,7 @@ module net is
   # 0  => arr_result[0] is the socket descriptor
   # -1 => arr_result[0] is an error number
   #
-  connect0(family, socket_type, protocol i32, host Any, port Any, arr_result Any) i32 is intrinsic
+  connect0(family, socket_type, protocol i32, host fuzion.sys.Pointer, port fuzion.sys.Pointer, arr_result fuzion.sys.Pointer) i32 is intrinsic
 
 
   # open and connect a client socket
@@ -112,7 +112,7 @@ module net is
   # true  => arr_result[0] is the number of bytes read
   # false => arr_result[0] is an error number
   #
-  read(sd i64, arr_data Any, length i32, arr_result Any) bool is intrinsic
+  read(sd i64, arr_data fuzion.sys.Pointer, length i32, arr_result fuzion.sys.Pointer) bool is intrinsic
 
 
   # read a maximum of max_bytes from descriptor, wrapper for read intrinsic
@@ -135,7 +135,7 @@ module net is
   #
   # returns zero on success, anything else is an error.
   #
-  write(sd i64, arr_data Any, length i32) i32 is intrinsic
+  write(sd i64, arr_data fuzion.sys.Pointer, length i32) i32 is intrinsic
 
 
   # write data to descriptor, wrapper for write intrinsic
@@ -187,7 +187,7 @@ module net is
   # returns the length of the IP address written to the array,
   # in bytes (4 for IPv4, 16 for IPv6)
   #
-  get_peer_address(sockfd i64, address Any) i32 is intrinsic
+  get_peer_address(sockfd i64, address fuzion.sys.Pointer) i32 is intrinsic
 
 
   # get a socket's peer's ip address
