+++ src/dev/flang/be/jvm/Intrinsix.java
@@ -401,19 +401,6 @@ public class Intrinsix extends ANY implements ClassFileConstants
           return new Pair<>(res, Expr.UNIT);
         });
 
-    put("fuzion.std.date_time",
-        (jvm, cl, pre, cc, tvalue, args) ->
-        {
-          var res =
-            args.get(0)
-            .andThen(Expr.checkcast(PrimitiveType.type_int.array()))
-            .andThen(Expr.invokeStatic(Names.RUNTIME_CLASS,
-                                       "fuzion_std_date_time",
-                                       "([I)V",
-                                       PrimitiveType.type_void));
-          return new Pair<>(Expr.UNIT, res);
-        });
-
     put("fuzion.sys.args.count",
         (jvm, cl, pre, cc, tvalue, args) ->
         {
@@ -432,21 +419,6 @@ public class Intrinsix extends ANY implements ClassFileConstants
                                                             Names.RUNTIME_ARGS_GET,
                                                             Names.RUNTIME_ARGS_GET_SIG,
                                                             PrimitiveType.type_byte.array())));
-        });
-
-    put("fuzion.sys.fileio.write", (jvm, cl, pre, cc, tvalue, args) ->
-        {
-          var res =
-            tvalue.drop()
-            .andThen(args.get(0))
-            .andThen(args.get(1))
-            .andThen(Expr.checkcast(PrimitiveType.type_byte.array()))
-            .andThen(args.get(2))
-            .andThen(Expr.invokeStatic(Names.RUNTIME_CLASS,
-                                       "fuzion_sys_fileio_write",
-                                       "(J[BI)I",
-                                       PrimitiveType.type_int));
-          return new Pair<>(res, Expr.UNIT);
         });
 
     put("fuzion.sys.internal_array.get",
@@ -573,312 +545,6 @@ public class Intrinsix extends ANY implements ClassFileConstants
           return new Pair<>(Expr.UNIT, Expr.iconst(jvm._options.fuzionSafety() ? 1 : 0));
         });
 
-
-    put("fuzion.sys.net.accept",(jvm, cl, pre, cc, tvalue, args) ->
-    {
-          var res =
-            tvalue.drop()
-            .andThen(args.get(0))
-            .andThen(args.get(1))
-            .andThen(Expr.checkcast(PrimitiveType.type_long.array()))
-            .andThen(Expr.invokeStatic(Names.RUNTIME_CLASS,
-                                       "fuzion_sys_net_accept",
-                                       methodDescriptor(Runtime.class, "fuzion_sys_net_accept"),
-                                       PrimitiveType.type_boolean));
-          return new Pair<>(res, Expr.UNIT);
-    });
-    put("fuzion.sys.net.bind0",(jvm, cl, pre, cc, tvalue, args) ->
-    {
-          var res =
-            tvalue.drop()
-            .andThen(args.get(0))
-            .andThen(args.get(1))
-            .andThen(args.get(2))
-            .andThen(args.get(3))
-            .andThen(Expr.checkcast(PrimitiveType.type_byte.array()))
-            .andThen(args.get(4))
-            .andThen(Expr.checkcast(PrimitiveType.type_byte.array()))
-            .andThen(args.get(5))
-            .andThen(Expr.checkcast(PrimitiveType.type_long.array()))
-            .andThen(Expr.invokeStatic(Names.RUNTIME_CLASS,
-                                       "fuzion_sys_net_bind0",
-                                       methodDescriptor(Runtime.class, "fuzion_sys_net_bind0"),
-                                       PrimitiveType.type_int));
-          return new Pair<>(res, Expr.UNIT);
-    });
-    put("fuzion.sys.net.close0",(jvm, cl, pre, cc, tvalue, args) ->
-    {
-          var res =
-            tvalue.drop()
-            .andThen(args.get(0))
-            .andThen(Expr.invokeStatic(Names.RUNTIME_CLASS,
-                                       "fuzion_sys_net_close0",
-                                        methodDescriptor(Runtime.class, "fuzion_sys_net_close0"),
-                                       PrimitiveType.type_int));
-          return new Pair<>(res, Expr.UNIT);
-    });
-    put("fuzion.sys.net.connect0",(jvm, cl, pre, cc, tvalue, args) ->
-    {
-          var res =
-            tvalue.drop()
-            .andThen(args.get(0))
-            .andThen(args.get(1))
-            .andThen(args.get(2))
-            .andThen(args.get(3))
-            .andThen(Expr.checkcast(PrimitiveType.type_byte.array()))
-            .andThen(args.get(4))
-            .andThen(Expr.checkcast(PrimitiveType.type_byte.array()))
-            .andThen(args.get(5))
-            .andThen(Expr.checkcast(PrimitiveType.type_long.array()))
-            .andThen(Expr.invokeStatic(Names.RUNTIME_CLASS,
-                                       "fuzion_sys_net_connect0",
-                                       methodDescriptor(Runtime.class, "fuzion_sys_net_connect0"),
-                                       PrimitiveType.type_int));
-          return new Pair<>(res, Expr.UNIT);
-    });
-    put("fuzion.sys.net.get_peer_address",(jvm, cl, pre, cc, tvalue, args) ->
-    {
-          var res =
-            tvalue.drop()
-            .andThen(args.get(0))
-            .andThen(args.get(1))
-            .andThen(Expr.checkcast(PrimitiveType.type_byte.array()))
-            .andThen(Expr.invokeStatic(Names.RUNTIME_CLASS,
-                                       "fuzion_sys_net_get_peer_address",
-                                       methodDescriptor(Runtime.class, "fuzion_sys_net_get_peer_address"),
-                                       PrimitiveType.type_int));
-          return new Pair<>(res, Expr.UNIT);
-    });
-    put("fuzion.sys.net.get_peer_port",(jvm, cl, pre, cc, tvalue, args) ->
-    {
-          var res =
-            tvalue.drop()
-            .andThen(args.get(0))
-            .andThen(Expr.invokeStatic(Names.RUNTIME_CLASS,
-                                       "fuzion_sys_net_get_peer_port",
-                                       methodDescriptor(Runtime.class, "fuzion_sys_net_get_peer_port"),
-                                       PrimitiveType.type_short));
-          return new Pair<>(res, Expr.UNIT);
-    });
-    put("fuzion.sys.net.listen",(jvm, cl, pre, cc, tvalue, args) ->
-    {
-          var res =
-            tvalue.drop()
-            .andThen(Expr.invokeStatic(Names.RUNTIME_CLASS,
-                                       "fuzion_sys_net_listen",
-                                       methodDescriptor(Runtime.class, "fuzion_sys_net_listen"),
-                                       PrimitiveType.type_int));
-          return new Pair<>(res, Expr.UNIT);
-    });
-    put("fuzion.sys.net.read",(jvm, cl, pre, cc, tvalue, args) ->
-    {
-          var res =
-            tvalue.drop()
-            .andThen(args.get(0))
-            .andThen(args.get(1))
-            .andThen(Expr.checkcast(PrimitiveType.type_byte.array()))
-            .andThen(args.get(2))
-            .andThen(args.get(3))
-            .andThen(Expr.checkcast(PrimitiveType.type_long.array()))
-            .andThen(Expr.invokeStatic(Names.RUNTIME_CLASS,
-                                       "fuzion_sys_net_read",
-                                       methodDescriptor(Runtime.class, "fuzion_sys_net_read"),
-                                       PrimitiveType.type_boolean));
-          return new Pair<>(res, Expr.UNIT);
-    });
-    put("fuzion.sys.net.set_blocking0",(jvm, cl, pre, cc, tvalue, args) ->
-    {
-          var res =
-            tvalue.drop()
-            .andThen(args.get(0))
-            .andThen(args.get(1))
-            .andThen(Expr.invokeStatic(Names.RUNTIME_CLASS,
-                                       "fuzion_sys_net_set_blocking0",
-                                       methodDescriptor(Runtime.class, "fuzion_sys_net_set_blocking0"),
-                                       PrimitiveType.type_int));
-          return new Pair<>(res, Expr.UNIT);
-    });
-    put("fuzion.sys.net.write",(jvm, cl, pre, cc, tvalue, args) ->
-    {
-          var res =
-            tvalue.drop()
-            .andThen(args.get(0))
-            .andThen(args.get(1))
-            .andThen(Expr.checkcast(PrimitiveType.type_byte.array()))
-            .andThen(Expr.invokeStatic(Names.RUNTIME_CLASS,
-                                       "fuzion_sys_net_write",
-                                       methodDescriptor(Runtime.class, "fuzion_sys_net_write"),
-                                       PrimitiveType.type_int));
-          return new Pair<>(res, Expr.UNIT);
-    });
-
-
-    put("fuzion.sys.fileio.close", (jvm, cl, pre, cc, tvalue, args) -> {
-      var res =
-        tvalue.drop()
-          .andThen(args.get(0))
-          .andThen(Expr.invokeStatic(Names.RUNTIME_CLASS,
-            "fuzion_sys_fileio_close",
-            methodDescriptor(Runtime.class, "fuzion_sys_fileio_close"),
-            PrimitiveType.type_int));
-      return new Pair<>(res, Expr.UNIT);
-    });
-    put("fuzion.sys.fileio.create_dir", (jvm, cl, pre, cc, tvalue, args) -> {
-      var res =
-        tvalue.drop()
-          .andThen(args.get(0))
-          .andThen(Expr.checkcast(PrimitiveType.type_byte.array()))
-          .andThen(Expr.invokeStatic(Names.RUNTIME_CLASS,
-            "fuzion_sys_fileio_create_dir",
-            methodDescriptor(Runtime.class, "fuzion_sys_fileio_create_dir"),
-            PrimitiveType.type_boolean));
-      return new Pair<>(res, Expr.UNIT);
-    });
-    put("fuzion.sys.fileio.delete", (jvm, cl, pre, cc, tvalue, args) -> {
-      var res =
-        tvalue.drop()
-          .andThen(args.get(0))
-          .andThen(Expr.checkcast(PrimitiveType.type_byte.array()))
-          .andThen(Expr.invokeStatic(Names.RUNTIME_CLASS,
-            "fuzion_sys_fileio_delete",
-            methodDescriptor(Runtime.class, "fuzion_sys_fileio_delete"),
-            PrimitiveType.type_boolean));
-      return new Pair<>(res, Expr.UNIT);
-    });
-    put("fuzion.sys.fileio.file_position", (jvm, cl, pre, cc, tvalue, args) -> {
-      var res =
-        tvalue.drop()
-          .andThen(args.get(0))
-          .andThen(args.get(1))
-          .andThen(Expr.checkcast(PrimitiveType.type_long.array()))
-          .andThen(Expr.invokeStatic(Names.RUNTIME_CLASS,
-            "fuzion_sys_fileio_file_position",
-            methodDescriptor(Runtime.class, "fuzion_sys_fileio_file_position"),
-            PrimitiveType.type_void));
-      return new Pair<>(Expr.UNIT, res);
-    });
-    put("fuzion.sys.fileio.flush", (jvm, cl, pre, cc, tvalue, args) -> {
-      var res =
-        tvalue.drop()
-          .andThen(args.get(0))
-          .andThen(Expr.invokeStatic(Names.RUNTIME_CLASS,
-            "fuzion_sys_fileio_flush",
-            methodDescriptor(Runtime.class, "fuzion_sys_fileio_flush"),
-            PrimitiveType.type_int));
-      return new Pair<>(res, Expr.UNIT);
-    });
-    put("fuzion.sys.fileio.lstats", (jvm, cl, pre, cc, tvalue, args) -> {
-      var res =
-        tvalue.drop()
-          .andThen(args.get(0))
-          .andThen(Expr.checkcast(PrimitiveType.type_byte.array()))
-          .andThen(args.get(1))
-          .andThen(Expr.checkcast(PrimitiveType.type_long.array()))
-          .andThen(Expr.invokeStatic(Names.RUNTIME_CLASS,
-            "fuzion_sys_fileio_lstats",
-            methodDescriptor(Runtime.class, "fuzion_sys_fileio_lstats"),
-            PrimitiveType.type_boolean));
-      return new Pair<>(res, Expr.UNIT);
-    });
-    put("fuzion.sys.fileio.stats", (jvm, cl, pre, cc, tvalue, args) -> {
-      var res =
-        tvalue.drop()
-          .andThen(args.get(0))
-          .andThen(Expr.checkcast(PrimitiveType.type_byte.array()))
-          .andThen(args.get(1))
-          .andThen(Expr.checkcast(PrimitiveType.type_long.array()))
-          .andThen(Expr.invokeStatic(Names.RUNTIME_CLASS,
-            "fuzion_sys_fileio_stats",
-            methodDescriptor(Runtime.class, "fuzion_sys_fileio_stats"),
-            PrimitiveType.type_boolean));
-      return new Pair<>(res, Expr.UNIT);
-    });
-    put("fuzion.sys.fileio.mmap", (jvm, cl, pre, cc, tvalue, args) -> {
-      var res =
-        tvalue.drop()
-          .andThen(args.get(0))
-          .andThen(args.get(1))
-          .andThen(args.get(2))
-          .andThen(args.get(3))
-          .andThen(Expr.checkcast(PrimitiveType.type_int.array()))
-          .andThen(Expr.invokeStatic(Names.RUNTIME_CLASS,
-            "fuzion_sys_fileio_mmap",
-            methodDescriptor(Runtime.class, "fuzion_sys_fileio_mmap"),
-            PrimitiveType.type_byte.array()));
-      return new Pair<>(res, Expr.UNIT);
-    });
-    put("fuzion.sys.fileio.move", (jvm, cl, pre, cc, tvalue, args) -> {
-      var res =
-        tvalue.drop()
-          .andThen(args.get(0))
-          .andThen(Expr.checkcast(PrimitiveType.type_byte.array()))
-          .andThen(args.get(1))
-          .andThen(Expr.checkcast(PrimitiveType.type_byte.array()))
-          .andThen(Expr.invokeStatic(Names.RUNTIME_CLASS,
-            "fuzion_sys_fileio_move",
-            methodDescriptor(Runtime.class, "fuzion_sys_fileio_move"),
-            PrimitiveType.type_boolean));
-      return new Pair<>(res, Expr.UNIT);
-    });
-    put("fuzion.sys.fileio.munmap", (jvm, cl, pre, cc, tvalue, args) -> {
-      var res =
-        tvalue.drop()
-          .andThen(Expr.invokeStatic(Names.RUNTIME_CLASS,
-            "fuzion_sys_fileio_munmap",
-            methodDescriptor(Runtime.class, "fuzion_sys_fileio_munmap"),
-            PrimitiveType.type_int));
-      return new Pair<>(res, Expr.UNIT);
-    });
-    put("fuzion.sys.fileio.open", (jvm, cl, pre, cc, tvalue, args) -> {
-      var res =
-        tvalue.drop()
-          .andThen(args.get(0))
-          .andThen(Expr.checkcast(PrimitiveType.type_byte.array()))
-          .andThen(args.get(1))
-          .andThen(Expr.checkcast(PrimitiveType.type_long.array()))
-          .andThen(args.get(2))
-          .andThen(Expr.invokeStatic(Names.RUNTIME_CLASS,
-            "fuzion_sys_fileio_open",
-            methodDescriptor(Runtime.class, "fuzion_sys_fileio_open"),
-            PrimitiveType.type_void));
-      return new Pair<>(Expr.UNIT, res);
-    });
-    put("fuzion.sys.fileio.read", (jvm, cl, pre, cc, tvalue, args) -> {
-      var res =
-        tvalue.drop()
-          .andThen(args.get(0))
-          .andThen(args.get(1))
-          .andThen(Expr.checkcast(PrimitiveType.type_byte.array()))
-          .andThen(Expr.invokeStatic(Names.RUNTIME_CLASS,
-            "fuzion_sys_fileio_read",
-            methodDescriptor(Runtime.class, "fuzion_sys_fileio_read"),
-            PrimitiveType.type_int));
-      return new Pair<>(res, Expr.UNIT);
-    });
-    put("fuzion.sys.fileio.seek", (jvm, cl, pre, cc, tvalue, args) -> {
-      var res =
-        tvalue.drop()
-          .andThen(args.get(0))
-          .andThen(args.get(1))
-          .andThen(args.get(2))
-          .andThen(Expr.checkcast(PrimitiveType.type_long.array()))
-          .andThen(Expr.invokeStatic(Names.RUNTIME_CLASS,
-            "fuzion_sys_fileio_seek",
-            methodDescriptor(Runtime.class, "fuzion_sys_fileio_seek"),
-            PrimitiveType.type_void));
-      return new Pair<>(Expr.UNIT, res);
-    });
-
-    put("fuzion.std.nano_sleep", (jvm, cl, pre, cc, tvalue, args) -> {
-      var res =
-        tvalue.drop()
-          .andThen(args.get(0))
-          .andThen(Expr.invokeStatic(Names.RUNTIME_CLASS,
-            "fuzion_std_nano_sleep",
-            methodDescriptor(Runtime.class, "fuzion_std_nano_sleep"),
-            PrimitiveType.type_void));
-      return new Pair<>(Expr.UNIT, res);
-    });
     put("fuzion.std.nano_time", (jvm, cl, pre, cc, tvalue, args) -> {
       var res =
         tvalue.drop()
@@ -886,25 +552,6 @@ public class Intrinsix extends ANY implements ClassFileConstants
             "nanoTime",
             methodDescriptor(System.class, "nanoTime"),
             PrimitiveType.type_long));
-      return new Pair<>(res, Expr.UNIT);
-    });
-    put("fuzion.sys.env_vars.get0", (jvm, cl, pre, cc, tvalue, args) -> {
-      return jvm.constString(
-        tvalue.drop()
-          .andThen(args.get(0))
-          .andThen(Expr.invokeStatic(Names.RUNTIME_CLASS,
-            "fuzion_sys_env_vars_get0",
-            methodDescriptor(Runtime.class, "fuzion_sys_env_vars_get0"),
-            JAVA_LANG_STRING)));
-    });
-    put("fuzion.sys.env_vars.has0", (jvm, cl, pre, cc, tvalue, args) -> {
-      var res =
-        tvalue.drop()
-          .andThen(args.get(0))
-          .andThen(Expr.invokeStatic(Names.RUNTIME_CLASS,
-            "fuzion_sys_env_vars_has0",
-            methodDescriptor(Runtime.class, "fuzion_sys_env_vars_has0"),
-            PrimitiveType.type_boolean));
       return new Pair<>(res, Expr.UNIT);
     });
     put("fuzion.sys.env_vars.set0", (jvm, cl, pre, cc, tvalue, args) -> {
