+++ src/dev/flang/be/jvm/runtime/Runtime.java
@@ -30,37 +30,10 @@ import dev.flang.be.interpreter.OpenResources; // NYI: remove dependency!
 
 import dev.flang.util.ANY;
 import dev.flang.util.Errors;
-import dev.flang.util.List;
-
-import java.io.IOException;
-import java.io.OutputStream;
-import java.io.PrintWriter;
-import java.io.RandomAccessFile;
 import java.io.StringWriter;
-import java.io.PrintStream;
-import java.io.InputStream;
-
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
-import java.net.BindException;
-import java.net.InetSocketAddress;
-import java.nio.ByteBuffer;
-import java.nio.channels.ByteChannel;
-import java.nio.channels.DatagramChannel;
-import java.nio.channels.ServerSocketChannel;
-import java.nio.channels.SocketChannel;
-import java.nio.channels.FileChannel.MapMode;
-import java.nio.channels.spi.AbstractSelectableChannel;
 import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.attribute.BasicFileAttributes;
-import java.util.concurrent.TimeUnit;
-
-import java.util.Calendar;
-import java.util.Date;
-import java.util.GregorianCalendar;
-import java.util.TimeZone;
 
 
 /**
@@ -143,20 +116,6 @@ public class Runtime extends ANY
 
 
   /**
-   * Create a Java string from 0-terminated given byte array.
-   */
-  private static String utf8ByteArrayDataToString(byte[] ba)
-  {
-    var l = 0;
-    while (l < ba.length && ba[l] != 0)
-      {
-        l++;
-      }
-    return new String(ba, 0, l, StandardCharsets.UTF_8);
-  }
-
-
-  /**
    * This contains all open files/streams.
    */
   static OpenResources<AutoCloseable> _openStreams_ = new OpenResources<AutoCloseable>()
@@ -207,6 +166,20 @@ public class Runtime extends ANY
 
 
   /*-------------------------  static methods  --------------------------*/
+
+
+  /**
+   * Create a Java string from 0-terminated given byte array.
+   */
+  public static String utf8ByteArrayDataToString(byte[] ba)
+  {
+    var l = 0;
+    while (l < ba.length && ba[l] != 0)
+      {
+        l++;
+      }
+    return new String(ba, 0, l, StandardCharsets.UTF_8);
+  }
 
 
   /**
@@ -635,538 +608,6 @@ public class Runtime extends ANY
     Errors.fatal("CONTRACT FAILED: " + msg, stacktrace.toString());
   }
 
-
-  public static void fuzion_std_date_time(int[] arg0)
-  {
-    var date = new Date();
-    var calendar = new GregorianCalendar(TimeZone.getTimeZone("UTC"));
-    calendar.setTime(date);
-    arg0[0] = calendar.get(Calendar.YEAR);
-    arg0[1] = calendar.get(Calendar.DAY_OF_YEAR);
-    arg0[2] = calendar.get(Calendar.HOUR_OF_DAY);
-    arg0[3] = calendar.get(Calendar.MINUTE);
-    arg0[4] = calendar.get(Calendar.SECOND);
-    arg0[5] = calendar.get(Calendar.MILLISECOND) * 1000;
-  }
-
-
-  public static int fuzion_sys_fileio_write(long f, byte[] fileContent, int l)
-  {
-    try
-      {
-        var s = Runtime._openStreams_.get(f);
-        if (s instanceof RandomAccessFile raf)
-          {
-            /* NYI:
-            if (!ENABLE_UNSAFE_INTRINSICS)
-              {
-                Errors.fatal("*** error: unsafe feature "+innerClazz+" disabled");
-              }
-            */
-            raf.write(fileContent);
-          }
-        else if (s instanceof OutputStream os)
-          {
-            os.write(fileContent);
-          }
-        return 0;
-      }
-    catch (IOException e)
-      {
-        return -1;
-      }
-  }
-
-
-  public static int fuzion_sys_net_bind0(int family, int socketType, int protocol, byte[] host0, byte[] port0, long[] result)
-  {
-    var host = utf8ByteArrayDataToString(host0);
-    var port = utf8ByteArrayDataToString(port0);
-
-    if (family != 2 && family != 10)
-      {
-        throw new RuntimeException("NYI");
-      }
-    try
-      {
-        return switch (protocol)
-          {
-            case 6 ->
-              {
-                var ss = ServerSocketChannel.open();
-                ss.bind(new InetSocketAddress(host, Integer.parseInt(port)));
-                result[0] = Runtime._openStreams_.add(ss);
-                yield 0;
-              }
-            case 17 ->
-              {
-                var ss = DatagramChannel.open();
-                ss.bind(new InetSocketAddress(host, Integer.parseInt(port)));
-                result[0] = Runtime._openStreams_.add(ss);
-                yield 0;
-              }
-            default -> throw new RuntimeException("NYI");
-          };
-      }
-    catch(BindException e)
-      {
-        result[0] = SystemErrNo.EADDRINUSE.errno;
-        return -1;
-      }
-    catch(IOException e)
-      {
-        result[0] = -1;
-        return -1;
-      }
-  }
-
-  public static int fuzion_sys_net_listen()
-  {
-    return 0;
-  }
-
-  public static boolean fuzion_sys_net_accept(long sockfd, long[] result)
-  {
-    try
-      {
-        var asc = Runtime._openStreams_.get(sockfd);
-        if(asc instanceof ServerSocketChannel ssc)
-          {
-            var socket = ssc.accept();
-            result[0] = Runtime._openStreams_.add(socket);
-            return true;
-          }
-        else if(asc instanceof DatagramChannel dc)
-          {
-            result[0] = sockfd;
-            return true;
-          }
-        throw new RuntimeException("NYI");
-      }
-    catch(IOException e)
-      {
-        return false;
-      }
-  }
-
-  public static int fuzion_sys_net_connect0(int family, int socketType, int protocol, byte[] host0, byte[] port0, long[] result)
-  {
-    var host = utf8ByteArrayDataToString(host0);
-    var port = utf8ByteArrayDataToString(port0);
-    if (family != 2 && family != 10)
-      {
-        throw new RuntimeException("NYI");
-      }
-    try
-      {
-        return switch (protocol)
-          {
-            case 6 ->
-              {
-                var socket = SocketChannel.open();
-                socket.connect(new InetSocketAddress(host, Integer.parseInt(port)));
-                result[0] = Runtime._openStreams_.add(socket);
-                yield 0;
-              }
-            case 17 ->
-              {
-                var ss = DatagramChannel.open();
-                ss.connect(new InetSocketAddress(host, Integer.parseInt(port)));
-                result[0] = Runtime._openStreams_.add(ss);
-                yield 0;
-              }
-            default -> throw new RuntimeException("NYI");
-          };
-      }
-    catch(IOException e)
-      {
-        result[0] = SystemErrNo.ECONNREFUSED.errno;
-        return -1;
-      }
-  }
-
-  public static int fuzion_sys_net_get_peer_address(long sockfd, byte[] data)
-  {
-    try
-      {
-        if (Runtime._openStreams_.get(sockfd) instanceof SocketChannel sc)
-          {
-            byte[] address = ((InetSocketAddress)sc.getRemoteAddress()).getAddress().getAddress();
-            System.arraycopy(address, 0, data, 0, address.length);
-            return address.length;
-          }
-        return -1;
-      }
-    catch (IOException e)
-      {
-        return -1;
-      }
-  }
-
-  public static short fuzion_sys_net_get_peer_port(long sockfd)
-  {
-    try
-      {
-        if (Runtime._openStreams_.get(sockfd) instanceof SocketChannel sc)
-          {
-            return (short) ((InetSocketAddress)sc.getRemoteAddress()).getPort();
-          }
-        return 0;
-      }
-    catch (IOException e)
-      {
-        return 0;
-      }
-  }
-
-  public static boolean fuzion_sys_net_read(long sockfd, byte[] buff, int length, long[] result)
-  {
-    try
-      {
-        var desc = Runtime._openStreams_.get(sockfd);
-        // NYI blocking / none blocking read
-        long bytesRead;
-        if (desc instanceof DatagramChannel dc)
-          {
-            bytesRead = dc.receive(ByteBuffer.wrap(buff)) == null
-              ? 0
-              // NYI how to determine datagram length?
-              : buff.length;
-          }
-        else if (desc instanceof ByteChannel sc)
-          {
-            bytesRead = sc.read(ByteBuffer.wrap(buff));
-          }
-        else
-          {
-            throw new RuntimeException("NYI");
-          }
-        result[0] = bytesRead;
-        return bytesRead != -1;
-      }
-    catch(IOException e) //SocketTimeoutException and others
-      {
-        // unspecified error
-        result[0] = -1;
-        return false;
-      }
-  }
-
-  public static int fuzion_sys_net_write(long sockfd, byte[] fileContent)
-  {
-    try
-      {
-        var sc = (ByteChannel)Runtime._openStreams_.get(sockfd);
-        sc.write(ByteBuffer.wrap(fileContent));
-        return 0;
-      }
-    catch(IOException e)
-      {
-        return -1;
-      }
-  }
-
-  public static int fuzion_sys_net_close0(long sockfd)
-  {
-    return Runtime._openStreams_.remove(sockfd)
-      ? 0
-      : -1;
-  }
-
-  public static int fuzion_sys_net_set_blocking0(long sockfd, int blocking)
-  {
-    var asc = (AbstractSelectableChannel)Runtime._openStreams_.get(sockfd);
-    try
-      {
-        asc.configureBlocking(blocking == 1);
-        return 0;
-      }
-    catch(IOException e)
-      {
-        return -1;
-      }
-  }
-
-  public static int fuzion_sys_fileio_flush(long fd)
-  {
-    var s = _openStreams_.get(fd);
-    if (s instanceof PrintStream ps)
-      {
-        ps.flush();
-      }
-    return 0;
-  }
-
-  public static int fuzion_sys_fileio_read(long fd, byte[] byteArr)
-  {
-    try
-      {
-        var s = _openStreams_.get(fd);
-        int bytesRead = 0;
-        if (s instanceof RandomAccessFile raf)
-          {
-            bytesRead = raf.read(byteArr);
-          }
-        else
-          {
-            bytesRead = ((InputStream) s).read(byteArr);
-          }
-
-        return bytesRead;
-      }
-    catch (Exception e)
-      {
-        return -2;
-      }
-  }
-
-  public static int fuzion_sys_fileio_write(long fd, byte[] fileContent)
-  {
-    try
-      {
-        var s = _openStreams_.get(fd);
-        if (s instanceof RandomAccessFile raf)
-          {
-            raf.write(fileContent);
-          }
-        else
-          {
-            ((OutputStream) s).write(fileContent);
-          }
-        return 0;
-      }
-    catch (Exception e)
-      {
-        return -1;
-      }
-  }
-
-  public static boolean fuzion_sys_fileio_delete(byte[] file)
-  {
-    Path path = Path.of(utf8ByteArrayDataToString(file));
-    try
-      {
-        return Files.deleteIfExists(path);
-      }
-    catch (Exception e)
-      {
-        return false;
-      }
-  }
-
-  public static boolean fuzion_sys_fileio_move(byte[] o, byte[] n)
-  {
-    Path oldPath = Path.of(utf8ByteArrayDataToString(o));
-    Path newPath = Path.of(utf8ByteArrayDataToString(n));
-    try
-      {
-        Files.move(oldPath, newPath);
-        return true;
-      }
-    catch (Exception e)
-      {
-        return false;
-      }
-  }
-
-  public static boolean fuzion_sys_fileio_create_dir(byte[] d)
-  {
-    Path path = Path.of(utf8ByteArrayDataToString(d));
-    try
-      {
-        Files.createDirectory(path);
-        return true;
-      }
-    catch (Exception e)
-      {
-        return false;
-      }
-  }
-
-  public static void fuzion_sys_fileio_open(byte[] path, long[] open_results, short mode)
-  {
-    try
-      {
-        switch (mode)
-          {
-          case 0 :
-            RandomAccessFile fis = new RandomAccessFile(utf8ByteArrayDataToString(path), "r");
-            open_results[0] = _openStreams_.add(fis);
-            break;
-          case 1 :
-            RandomAccessFile fos = new RandomAccessFile(utf8ByteArrayDataToString(path), "rw");
-            open_results[0] = _openStreams_.add(fos);
-            break;
-          case 2 :
-            RandomAccessFile fas = new RandomAccessFile(utf8ByteArrayDataToString(path), "rw");
-            fas.seek(fas.length());
-            open_results[0] = _openStreams_.add(fas);
-            break;
-          default:
-            open_results[1] = -1;
-            System.err.println("*** Unsupported open flag. Please use: 0 for READ, 1 for WRITE, 2 for APPEND. ***");
-            System.exit(1);
-          }
-      }
-    catch (Exception e)
-      {
-        open_results[1] = -1;
-      }
-    return;
-  }
-
-  public static int fuzion_sys_fileio_close(long fd)
-  {
-    return _openStreams_.remove(fd)
-                                    ? 0
-                                    : -1;
-  }
-
-  public static boolean fuzion_sys_fileio_lstats(byte[] d, long[] stats) // NYI : should be altered in the future to not resolve symbolic links
-  {
-    return fuzion_sys_fileio_stats(d, stats);
-  }
-  public static boolean fuzion_sys_fileio_stats(byte[] d, long[] stats)
-  {
-    Path path = Path.of(utf8ByteArrayDataToString(d));
-    var err = SystemErrNo.UNSPECIFIED;
-    try
-      {
-        BasicFileAttributes metadata = Files.readAttributes(path, BasicFileAttributes.class);
-        stats[0] = metadata.size();
-        stats[1] = metadata.lastModifiedTime().to(TimeUnit.SECONDS);
-        stats[2] = metadata.isRegularFile() ? 1: 0;
-        stats[3] = metadata.isDirectory() ? 1: 0;
-        return true;
-      }
-    catch (UnsupportedOperationException e)
-      {
-        err = SystemErrNo.ENOTSUP;
-      }
-    catch (IOException e)
-      {
-        err = SystemErrNo.EIO;
-      }
-    catch (SecurityException e)
-      {
-        err = SystemErrNo.EACCES;
-      }
-
-    stats[0] = err.errno;
-    stats[1] = 0;
-    stats[2] = 0;
-    stats[3] = 0;
-    return false;
-  }
-
-  public static void fuzion_sys_fileio_seek(long fd, short s, long[] seekResults)
-  {
-    try
-      {
-        var raf = (RandomAccessFile) _openStreams_.get(fd);
-        raf.seek(s);
-        seekResults[0] = raf.getFilePointer();
-        return;
-      }
-    catch (Exception e)
-      {
-        seekResults[1] = -1;
-        return;
-      }
-  }
-
-  public static void fuzion_sys_fileio_file_position(long fd, long[] arr)
-  {
-    try
-      {
-        arr[0] = ((RandomAccessFile) _openStreams_.get(fd)).getFilePointer();
-        return;
-      }
-    catch (Exception e)
-      {
-        arr[1] = -1;
-        return;
-      }
-  }
-
-  public static byte[] fuzion_sys_fileio_mmap(long fd, long offset, long size, int[] result)
-  {
-    try
-      {
-        var raf = (RandomAccessFile) _openStreams_.get(fd);
-
-        // offset+size must not exceed file size, to match semantics of
-        // c-backend.
-        if (raf.length() < (offset + size))
-          {
-            result[0] = -1;
-            return new byte[0];
-          }
-
-        var mmap = raf.getChannel().map(MapMode.READ_WRITE, offset, size);
-
-        // success
-        result[0] = 0;
-        return new byte[0];
-        /* NYI
-                  @Override
-                  void set(
-                    int x,
-                    Value v,
-                    AbstractType elementType)
-                  {
-                    checkIndex(x);
-                    mmap.put(x, (byte)v.u8Value());
-                  }
-
-                  @Override
-                  Value get(
-                    int x,
-                    AbstractType elementType)
-                  {
-                    checkIndex(x);
-                    return new u8Value(mmap.get(x));
-                  }
-
-                  @Override
-                  int length(){
-                    return (int)size;
-                  }
-        */
-      }
-    catch (IOException e)
-      {
-        result[0] = -1;
-        return new byte[0];
-      }
-  }
-
-  public static int fuzion_sys_fileio_munmap()
-  {
-    return 0;
-  }
-
-  public static void fuzion_std_nano_sleep(long d)
-  {
-    try
-      {
-        TimeUnit.NANOSECONDS.sleep(d < 0 ? Long.MAX_VALUE: d);
-      }
-    catch (InterruptedException ie)
-      {
-        throw new Error("unexpected interrupt", ie);
-      }
-  };
-
-  public static boolean fuzion_sys_env_vars_has0(byte[] b)
-  {
-    return System.getenv(utf8ByteArrayDataToString(b)) != null;
-  }
-
-  public static String fuzion_sys_env_vars_get0(byte[] d)
-  {
-    return System.getenv(utf8ByteArrayDataToString(d));
-  }
 
 
   /**
